Initializing a support module to handle testing datasets
