+++ tests/test_IO.py
@@ -6,22 +6,7 @@ import matplotlib.pyplot as plt
 import pytest
 
 import lidarSuit as lst
-
-
-def get_sample_data(sample_path, file_type):
-
-    if file_type == "12-00":
-        url = "https://drive.google.com/uc?export=download&id=1i6iX6KuZOkP_WLuPZHG5uCcvRjlWS-SU"
-
-    if file_type == "dbs":
-        url = "path"
-
-    output = f"{sample_path}{file_type}.zip"
-    gdown.download(url, output, quiet=False)
-
-    print(f"Extracting: {output}")
-    shutil.unpack_archive(output, sample_path)
-    os.remove(output)
+from .data import get_sample_data
 
 
 @pytest.fixture
